requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pytesseract==0.3.10
sqlparse==0.5.0
tesseract==0.1.3
typing_extensions==4.12.2
tzdata==2024.1
